refactor: 캐스퍼 레이싱 섹션 컴포넌트 분리
