+++ packages/user/src/components/event/racing/controls/Gauge.tsx
